(refactor) moved attendanceHelpers into helpers folder
